WIP: Strip the token name out of certutil output
This will hopefully be merged into the first patch but the basic
jist is that we don't care about the token a certificate is stored
in, the internal or the FIPS token. We just care that they exist
on both servers and that the keys match.

Apparently in some cases the token name is displayed and not in
others so lets normalize the output to make comparisons more
consistent.